Rename panel files, cleanup test print statements
