+++ scripts/addons/cam/tests/gcode_generator.py
@@ -5,15 +5,16 @@ import bpy
 
 warnings.simplefilter("once")
 
-# Get the scene
-s = bpy.context.scene
-s.render.engine = "FABEX_RENDER"
+# Set the Render Engine to Fabex
+bpy.context.scene.render.engine = "FABEX_RENDER"
 
 for i, operation in enumerate(s.cam_operations):
     # Set the active operation using the index
     s.cam_active_operation = i
 
-    print(f"############ Generating {operation.name} #############")
+    print("")
+    print("")
+    print(f"############ Generating {operation.name} Operation ############")
 
     # Run the calculate_cam_path() operator
     bpy.ops.object.calculate_cam_path()
