+++ scripts/addons/cam/ui/__init__.py
@@ -16,26 +16,26 @@ from .menus.preset_menus import (
 )
 from .menus.viewport import Fabex_SubMenu, Fabex_Menu
 
-from .panels.area import CAM_AREA_Panel
-from .panels.blank import CAM_BLANK_Panel
-from .panels.chains import (
+from .panels.area_panel import CAM_AREA_Panel
+from .panels.blank_panel import CAM_BLANK_Panel
+from .panels.chains_panel import (
     CAM_CHAINS_Panel,
     CAM_UL_chains,
     CAM_UL_operations,
 )
-from .panels.curve_creators import VIEW3D_PT_tools_create
-from .panels.curve_tools import VIEW3D_PT_tools_curvetools
-from .panels.cutter import CAM_CUTTER_Panel
-from .panels.feedrate import CAM_FEEDRATE_Panel
-from .panels.gcode import CAM_GCODE_Panel
-from .panels.info import CAM_INFO_Panel
-from .panels.machine import CAM_MACHINE_Panel
-from .panels.material import CAM_MATERIAL_Panel
-from .panels.movement import CAM_MOVEMENT_Panel
-from .panels.op_properties import CAM_OPERATION_PROPERTIES_Panel
-from .panels.operations import CAM_OPERATIONS_Panel
-from .panels.optimisation import CAM_OPTIMISATION_Panel
-from .panels.popup import CAM_Popup_Panel
+from .panels.curve_create_panel import VIEW3D_PT_tools_create
+from .panels.curve_tools_panel import VIEW3D_PT_tools_curvetools
+from .panels.cutter_panel import CAM_CUTTER_Panel
+from .panels.feedrate_panel import CAM_FEEDRATE_Panel
+from .panels.gcode_panel import CAM_GCODE_Panel
+from .panels.info_panel import CAM_INFO_Panel
+from .panels.machine_panel import CAM_MACHINE_Panel
+from .panels.material_panel import CAM_MATERIAL_Panel
+from .panels.movement_panel import CAM_MOVEMENT_Panel
+from .panels.op_properties_panel import CAM_OPERATION_PROPERTIES_Panel
+from .panels.operations_panel import CAM_OPERATIONS_Panel
+from .panels.optimisation_panel import CAM_OPTIMISATION_Panel
+from .panels.popup_panel import CAM_Popup_Panel
 
 from .pie_menu.pie_cam import VIEW3D_MT_PIE_CAM
 from .pie_menu.pie_chains import VIEW3D_MT_PIE_Chains
