+++ scripts/addons/cam/ui/panels/area_panel.py
@@ -6,12 +6,12 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 # from ...utilities.simple_utils import unit_value_to_string
 
 
-class CAM_AREA_Panel(CAMButtonsPanel, Panel):
+class CAM_AREA_Panel(CAM_Parent_Panel, Panel):
     """CAM Operation Area Panel"""
 
     bl_space_type = "VIEW_3D"
