+++ scripts/addons/cam/ui/panels/blank_panel.py
