+++ scripts/addons/cam/ui/panels/curve_create_panel.py
