+++ scripts/addons/cam/ui/panels/curve_tools_panel.py
