+++ scripts/addons/cam/ui/panels/cutter_panel.py
@@ -6,10 +6,10 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 
-class CAM_CUTTER_Panel(CAMButtonsPanel, Panel):
+class CAM_CUTTER_Panel(CAM_Parent_Panel, Panel):
     """CAM Cutter Panel"""
 
     bl_space_type = "VIEW_3D"
