+++ scripts/addons/cam/ui/panels/feedrate_panel.py
@@ -6,10 +6,10 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 
-class CAM_FEEDRATE_Panel(CAMButtonsPanel, Panel):
+class CAM_FEEDRATE_Panel(CAM_Parent_Panel, Panel):
     """CAM Feedrate Panel"""
 
     bl_space_type = "VIEW_3D"
