+++ scripts/addons/cam/ui/panels/operations_panel.py
@@ -6,10 +6,10 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 
-class CAM_OPERATIONS_Panel(CAMButtonsPanel, Panel):
+class CAM_OPERATIONS_Panel(CAM_Parent_Panel, Panel):
     """CAM Operations Panel"""
 
     bl_space_type = "PROPERTIES"
