+++ scripts/addons/cam/ui/panels/pack.py
@@ -6,10 +6,10 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 
-class CAM_PACK_Panel(CAMButtonsPanel, Panel):
+class CAM_PACK_Panel(CAM_Parent_Panel, Panel):
     """CAM Pack Panel"""
 
     bl_space_type = "PROPERTIES"
