+++ scripts/addons/cam/ui/panels/slice.py
@@ -6,10 +6,10 @@
 import bpy
 from bpy.types import Panel
 
-from .buttons_panel import CAMButtonsPanel
+from .parent_panel import CAM_Parent_Panel
 
 
-class CAM_SLICE_Panel(CAMButtonsPanel, Panel):
+class CAM_SLICE_Panel(CAM_Parent_Panel, Panel):
     """CAM Slicer Panel"""
 
     bl_space_type = "PROPERTIES"
